+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 # 1.0.0
 ```
+    * In the admin added check for the merchant data before try to get merchant payment methods.
     * When call Checkokut SDK, pass the billing address and into userData block.
     * Removed nuveiLoadCheckout() method from frontend JS.
     * Removed nuveiGetSessionToken() method from frontend JS.